requirements/base.txt
asgiref==3.7.2
django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
django-environ==0.11.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
celery-progress==0.3
django_celery_beat==2.5.0
flower==2.0.1
requests==2.31.0